+++ eras/conway/impl/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## 1.12.0.0
 
-* Change 'getScriptWitnessConwayTxCert' so that DRepRegistration certificate requires a witness		
+* Change 'getScriptWitnessConwayTxCert' so that DRepRegistration certificate requires a witness
+* Implement `ToJSON` instance for `PoolVotingThresholds` and `DRepVotingThreshold`,
+  instead of deriving that doesn't handle field names correctly.
+* Hide `Cardano.Ledger.Conway.TxOut` module
+* Export `ConwayEraPParams` from `Cardano.Ledger.Conway.Core`
+* Stop exporting `BabbagePParams` from `Cardano.Ledger.Conway.PParams`
 * Add `transTxBodyWithdrawals`, `transTxCert`, `transDRepCred`, `transColdCommitteeCred`,
   `transHotCommitteeCred`, `transDelegatee`, `transDRep`, `transScriptPurpose`
 * Remove `conwayTxInfo` and `babbageScriptPrefixTag`
