+++ eras/conway/impl/cardano-ledger-conway.cabal
@@ -32,7 +32,6 @@ library
         Cardano.Ledger.Conway.Tx
         Cardano.Ledger.Conway.TxBody
         Cardano.Ledger.Conway.TxInfo
-        Cardano.Ledger.Conway.TxOut
         Cardano.Ledger.Conway.TxWits
         Cardano.Ledger.Conway.Transition
         Cardano.Ledger.Conway.Translation
@@ -63,6 +62,7 @@ library
         Cardano.Ledger.Conway.Rules.Utxos
         Cardano.Ledger.Conway.Rules.Utxow
         Cardano.Ledger.Conway.TxAuxData
+        Cardano.Ledger.Conway.TxOut
 
     default-language: Haskell2010
     ghc-options:
