+++ eras/conway/impl/test/Test/Cardano/Ledger/Conway/CommitteeRatifySpec.hs
@@ -22,7 +22,6 @@ import Cardano.Ledger.Conway.Governance (
   ensCommitteeL,
   rsEnactStateL,
  )
-import Cardano.Ledger.Conway.PParams (ConwayEraPParams)
 import Cardano.Ledger.Conway.Rules (
   committeeAccepted,
   committeeAcceptedRatio,
