+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EpochSpec.hs
@@ -21,10 +21,8 @@ import Cardano.Ledger.Conway.Governance (
   ProposalProcedure (..),
   Voter (..),
  )
-import Cardano.Ledger.Conway.PParams (
-  ppGovActionDepositL,
-  ppGovActionLifetimeL,
- )
+import Cardano.Ledger.Credential (Credential (..))
+import Cardano.Ledger.Keys
 import Cardano.Ledger.Shelley.LedgerState (
   asTreasuryL,
   esAccountStateL,
