+++ libs/cardano-ledger-conformance/cardano-ledger-conformance.cabal
@@ -40,6 +40,7 @@ library
         cardano-ledger-mary,
         cardano-ledger-shelley,
         cardano-ledger-alonzo,
+        cardano-ledger-babbage,
         cardano-ledger-conway:{cardano-ledger-conway, testlib},
         cardano-ledger-executable-spec,
         cardano-crypto-class,
